requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
google-cloud-texttospeech==2.16.3
google-api-core==2.30.3
pydantic==2.7.0
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv==1.0.1
redis==5.0.4
tenacity==8.3.0
prometheus-client==0.20.0
google-cloud-bigquery==3.21.0
